refactor(content): move compileBundle into src/lib/content/compile so route+script share one impl
compileBundle (pure, in-memory tree -> {files, assets}) and its
script-local helpers (validateTree, planAssets, validateAssetReferences,
stableJson, etc.) move from scripts/compile-content.ts into a new
src/lib/content/compile/compile-bundle.ts, alongside the projector/
validate/tarball helpers it already depended on. scripts/compile-content.ts
now imports and re-exports compileBundle/compileContent; main()/writeBundle()
and the GitHub tarball fetch stay in the script. No logic changed.
